# ranger: Improve intsRanger performance