@sauerbraten 2022-12-15
Very cool, I think the new intsRanger code is still pretty easy to read, so am definitely gonna merge this. Thank you!